src/requirements.txt
psycopg2-binary==2.9.4 
email-validator==1.3.0
fastapi==0.85.0
bcrypt==4.0.1
cryptography==38.0.2
passlib==1.7.4
pydantic==1.10.2
python-jose==3.3.0
python-multipart==0.0.5
SQLAlchemy==1.4.41
starlette==0.20.4
typing_extensions==4.4.0
uvicorn==0.18.3
websockets==10.3
